chore: Separate expander_task_list_utils and expander_task_utils
Also change `any_wildcard_task()` to `any_task_of_type()` for more general use.
